refactor all architectures except internal
