+++ README.md
@@ -38,7 +38,7 @@ TorchMultimodal requires Python >= 3.7. The library can be installed with or wit
 ## Documentation
 
 The library builds on the following concepts:
-- **Architectures**: These are general and composable classes that capture the core logic associated with a family of models. In most cases these take modules as inputs instead of flat arguments (see Models below). Examples include the `LateFusionArchitecture`, `FLAVA` and `CLIPArchitecture`. Users should either reuse an existing architecture or a contribute a new one. We avoid inheritance as much as possible.
+- **Architectures**: These are general and composable classes that capture the core logic associated with a family of models. In most cases these take modules as inputs instead of flat arguments (see Models below). Examples include the `LateFusion`, `FLAVA` and `CLIP`. Users should either reuse an existing architecture or a contribute a new one. We avoid inheritance as much as possible.
 
 - **Models**: These are specific instantiations of a given architecture implemented using builder functions. The builder functions take as input all of the parameters for constructing the modules needed to instantiate the architecture. See [cnn_lstm.py](https://github.com/facebookresearch/multimodal/blob/main/torchmultimodal/models/cnn_lstm.py) for an example.
