+++ torchmultimodal/models/clip.py
@@ -4,11 +4,62 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from torchmultimodal.architectures.clip import CLIPArchitecture
+
+from typing import NamedTuple
+
+import torch
+import torch.nn.functional as F
+from torch import nn
+
 from torchmultimodal.modules.encoders.clip_resnet_encoder import ResNetForCLIP
 from torchmultimodal.modules.encoders.clip_text_encoder import CLIPTextEncoder
 from torchvision.models.resnet import Bottleneck, ResNet
 from torchvision.models.vision_transformer import VisionTransformer
+
+
+class CLIPOutput(NamedTuple):
+    embeddings_a: torch.Tensor
+    embeddings_b: torch.Tensor
+
+
+class CLIP(nn.Module):
+    """CLIP is a model for contrastive pretraining between two modalities.
+
+    CLIP (https://arxiv.org/pdf/2103.00020.pdf) jointly trains an image encoder
+    (either ResNet or ViT) and a text encoder (Transformer) to predict correct
+    (image, text) pairings via a contrastive loss function. This module contains the
+    encoders, while the loss is implemented in ContrastiveLossWithTemperature.
+
+
+    Args:   encoder_a (nn.Module): Instantiated encoder for modality A.
+                See e.g. ResNetForCLIP class.
+            encoder_b (nn.Module): Instantiated encoder for modality B.
+                See e.g. CLIPTextEncoder class.
+
+    Inputs: features_a (Tensor): Tensor containing features of modality A.
+            features_b (Tensor): Tensor containing features of modality B.
+    """
+
+    def __init__(
+        self,
+        encoder_a: nn.Module,
+        encoder_b: nn.Module,
+    ):
+        super().__init__()
+        self.encoder_a = encoder_a
+        self.encoder_b = encoder_b
+
+    def forward(
+        self,
+        features_a: torch.Tensor,
+        features_b: torch.Tensor,
+    ) -> CLIPOutput:
+
+        embeddings_a = self.encoder_a(features_a)
+        embeddings_b = self.encoder_b(features_b)
+        embeddings_a = F.normalize(embeddings_a)
+        embeddings_b = F.normalize(embeddings_b)
+        return CLIPOutput(embeddings_a=embeddings_a, embeddings_b=embeddings_b)
 
 
 def clip_vit_b16():
@@ -22,7 +73,7 @@ def clip_vit_b16():
         num_classes=512,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=512)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_vit_b32():
@@ -36,7 +87,7 @@ def clip_vit_b32():
         num_classes=512,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=512)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_vit_l14():
@@ -50,7 +101,7 @@ def clip_vit_l14():
         num_classes=768,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=768, width=768, heads=12)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_rn50():
@@ -61,7 +112,7 @@ def clip_rn50():
         width=2048,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=1024)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_rn101():
@@ -72,7 +123,7 @@ def clip_rn101():
         width=2048,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=1024)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 # Note: these models require larger image sizes
@@ -85,7 +136,7 @@ def clip_rn50x4():
         width=2560,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=1024, width=640, heads=12)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_rn50x16():
@@ -97,7 +148,7 @@ def clip_rn50x16():
         width=3072,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=768, width=768, heads=12)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_rn50x64():
@@ -109,7 +160,7 @@ def clip_rn50x64():
         width=4096,
     )
     text_encoder = CLIPTextEncoder(embedding_dim=1024, width=1024, heads=16)
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 # Note: these models use torchvision's ResNet
@@ -120,7 +171,7 @@ def clip_rn50_tv():
         num_classes=1024,
     )
     text_encoder = CLIPTextEncoder()
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
 
 
 def clip_rn101_tv():
@@ -130,4 +181,4 @@ def clip_rn101_tv():
         num_classes=512,
     )
     text_encoder = CLIPTextEncoder()
-    return CLIPArchitecture(vision_encoder, text_encoder)
+    return CLIP(vision_encoder, text_encoder)
