+++ torchmultimodal/models/omnivore.py
@@ -9,7 +9,6 @@ from typing import Callable, List, Optional
 
 import torch
 from torch import nn
-from torchmultimodal.architectures.omnivore import OmnivoreArchitecture
 from torchmultimodal.modules.encoders.swin_transformer_3d_encoder import (
     PatchEmbed3d,
     SwinTransformer3d,
@@ -39,6 +38,36 @@ def _multimodal_head(input_dim: int) -> nn.ModuleDict:
             "video": _kinetics400_head(input_dim),
         }
     )
+
+
+class Omnivore(nn.Module):
+    """Omnivore is a model that accept multiple vision modality.
+
+    Omnivore (https://arxiv.org/abs/2201.08377) is a single model that able to do classification
+    on images, videos, and single-view 3D data using the same shared parameters of the encoder.
+
+    Args:   encoder (nn.Module): Instantiated encoder.
+                See SwinTransformer3dEncoder class.
+            heads (Optinal[nn.ModuleDict]): Dictionary of multiple heads for each dataset type
+
+    Inputs: x (Tensor): 5 Dimensional batched video tensor with format of B C D H W
+                where B is batch, C is channel, D is time, H is height, and W is width.
+            input_type (str): The dataset type of the input, this will used to choose
+                the correct head.
+    """
+
+    def __init__(self, encoder: nn.Module, heads: nn.ModuleDict):
+        super().__init__()
+        self.encoder = encoder
+        self.heads = heads
+
+    def forward(self, x: torch.Tensor, input_type: str):
+        x = self.encoder(x)
+        assert (
+            input_type in self.heads
+        ), f"Unsupported input_type: {input_type}, please use one of {list(self.heads.keys())}"
+        x = self.heads[input_type](x)
+        return x
 
 
 class PatchEmbedOmnivore(nn.Module):
@@ -111,4 +140,4 @@ def omnivore_swin_t(encoder_only=False) -> nn.Module:
         return encoder
 
     heads = _multimodal_head(input_dim=encoder.num_features)
-    return OmnivoreArchitecture(encoder, heads)
+    return Omnivore(encoder, heads)
